requirements.txt
Flask==3.0.3
gunicorn==21.2.0
Pillow==10.3.0
pandas==2.2.2
openpyxl==3.1.2
bcrypt==4.1.2
qrcode==7.4.2